requirements.txt
python-telegram-bot==20.5
yt-dlp>=2023.12.1
